+++ src/partialMatchRegExp/index.ts
@@ -1,0 +1,3 @@
+export * from "./partialMatchRegExp.ts";
+export { default } from "./partialMatchRegExp.ts";
+export { default as isComplete } from "./isComplete/index.ts";